fix: correctifs dashbaord + list operateur
