+++ src/pages/certification/exploitations/index.vue
@@ -727,7 +727,7 @@ function hideTooltip() {
 }
 
 watch(
-  [searchInput, currentPage, sortOrder, currentFilter.value],
+  [searchInput, currentPage, sortOrder.value, currentFilter.value],
   () => doSearch(searchInput.value, currentPage.value, sortOrder.value.sort, sortOrder.value.order),
   { immediate: true },
 );
