+++ src/pages/exploitations/[numeroBio]/index.vue
@@ -17,28 +17,28 @@ meta:
         </li>
       </ol>
     </nav>
-    <h1 class="fr-h3 fr-m-0">
-      {{ operatorStore.operator.nom }}
-    </h1>
-    <div class="header fr-mb-8v">
-      <NotificationState v-if="operatorStore.operator.notifications" :operator="operatorStore.operator" :text="true" />
-      <div v-else></div>
+    <div class="flex-center">
+      <h1 class="fr-h3 fr-m-0">
+        {{ operatorStore.operator.nom }}
+      </h1>
       <template v-if="permissions.isOc">
         <button
           v-if="operatorStore.operator.epingle"
-          class="fr-btn fr-btn--secondary fr-btn--icon-left fr-icon-add-line"
+          class="fr-btn fr-btn--tertiary-no-outline ri-pushpin-fill"
           @click="unpin(operatorStore.operator.numeroBio)"
-        >
-          Désepingler
-        </button>
+          aria-label="Désepingler"
+        ></button>
         <button
           v-else
-          class="fr-btn fr-btn--secondary fr-btn--icon-left fr-icon-add-line"
+          class="fr-btn fr-btn--tertiary-no-outline ri-pushpin-line"
           @click="pin(operatorStore.operator.numeroBio)"
-        >
-          Epingler
-        </button>
+          aria-label="Epingler"
+        ></button>
       </template>
+    </div>
+    <div class="header fr-mb-8v">
+      <NotificationState v-if="operatorStore.operator.notifications" :operator="operatorStore.operator" :text="true" />
+      <div v-else></div>
       <button
         v-if="permissions.canCreateVersion && operatorStore.records && operatorStore.records.length > 0"
         class="fr-btn fr-btn--secondary fr-btn--icon-left fr-icon-add-line create-new-version"
@@ -547,5 +547,10 @@ tr.version-row:hover {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
+}
+
+.flex-center {
+  display: flex;
+  align-items: center;
 }
 </style>
